requirements.txt
anthropic>=0.40.0
google-generativeai>=0.8.0
openai>=1.54.0
pyyaml>=6.0
aiohttp>=3.11.0
